ebb: Repaired Schematron Files!
I made some repairs to both Schematron files so they should work to do
what we want now!
1) In the ProsopCheck.sch, to make sure there any ref attributes have
hashtags (just added a rule), and
2) In the OHaraTextsCheck.sch to make sure that *ANY PORTION* of a
persName or placeName in the poetry file matches something inside a
persName or placeName in your prosopography file. You were getting
errors before for names like <persName>Miss Stillwagon</persName>,
because in your Prosopography entry for Linda Stillwagon, there was no
"Miss" inside that Prosop. <persName>. What we're doing is making it so
ANY PART of a persName that matches something in the appropriate Prosop
entry will pass our Schematron test.
3) One more change: I made a rule catch when a white space begins or
ends the text contents of a  persName or placeName element: We need to
prevent those white spaces if we want the matching of name parts with
the Prosopography file to work. (It's not really a problem in your files
from what I can see, but it's a "just-in-case" measure.
4) I deleted some older versions of these schematron files that we'd
generated last week. Make sure to associate your poems and prosop file
with the current schematron files, and report back if you see anything
weird happening again with validation! @rocketsaucer @RJP43